- Merge of Default and Issue Hierarchy Fatories to one class.
