+++ release_notes_generator/utils/constants.py
@@ -39,7 +39,7 @@ ROW_FORMAT_ISSUE = "row-format-issue"
 ROW_FORMAT_PR = "row-format-pr"
 ROW_FORMAT_LINK_PR = "row-format-link-pr"
 SUPPORTED_ROW_FORMAT_KEYS_HIERARCHY_ISSUE = ["type", "number", "title"]
-SUPPORTED_ROW_FORMAT_KEYS_ISSUE = ["number", "title", "pull-requests"]
+SUPPORTED_ROW_FORMAT_KEYS_ISSUE = ["type", "number", "title", "pull-requests"]
 SUPPORTED_ROW_FORMAT_KEYS_PULL_REQUEST = ["number", "title"]
 
 # Features
