+++ tests/release_notes/builder/test_release_notes_builder.py
@@ -20,7 +20,6 @@ from github import Github
 from release_notes_generator.builder.builder import ReleaseNotesBuilder
 from release_notes_generator.chapters.custom_chapters import CustomChapters
 from release_notes_generator.record.factory.default_record_factory import DefaultRecordFactory
-from release_notes_generator.record.factory.issue_hierarchy_record_factory import IssueHierarchyRecordFactory
 from tests.conftest import mock_safe_call_decorator, MockLabel
 
 # pylint: disable=pointless-string-statement
@@ -138,7 +137,7 @@ http://example.com/changelog
 RELEASE_NOTES_NO_DATA_NO_EMPTY_CHAPTERS = RELEASE_NOTES_NO_DATA_NO_WARNING_NO_EMPTY_CHAPTERS
 
 RELEASE_NOTES_DATA_CUSTOM_CHAPTERS_ONE_LABEL = """### Chapter 1 🛠
-- #122 _I1+bug_ in #101, #102
+- N/A: #122 _I1+bug_ in #101, #102
   - Fixed bug
   - Improved performance
   + More nice code
@@ -153,11 +152,11 @@ http://example.com/changelog
 """
 
 RELEASE_NOTES_DATA_HIERARCHY_NO_LABELS_NO_TYPE = """### Closed Issues without Pull Request ⚠️
-- #121 _Fix the bug_ in 
+- N/A: #121 _Fix the bug_ in 
   - Solo issue release note
 
 ### Closed Issues without User Defined Labels ⚠️
-- 🔔 #121 _Fix the bug_ in 
+- 🔔 N/A: #121 _Fix the bug_ in 
   - Solo issue release note
 
 ### Merged PRs without Issue and User Defined Labels ⚠️
@@ -176,7 +175,7 @@ RELEASE_NOTES_DATA_HIERARCHY_NO_LABELS_NO_TYPE = """### Closed Issues without Pu
 - None: _HI302 open_ #302
   - _Release Notes_:
     - Hierarchy level release note
-  - #451 _SI451 closed_ in #150
+  - N/A: #451 _SI451 closed_ in #150
     - Hierarchy level release note
     - Fixed bug
     - Improved performance
@@ -185,7 +184,7 @@ RELEASE_NOTES_DATA_HIERARCHY_NO_LABELS_NO_TYPE = """### Closed Issues without Pu
 - None: _HI303 open_ #303
   - _Release Notes_:
     - Hierarchy level release note
-  - #452 _SI452 closed_ in #151
+  - N/A: #452 _SI452 closed_ in #151
     - Hierarchy level release note
     - Fixed bug
     - Improved performance
@@ -197,7 +196,7 @@ RELEASE_NOTES_DATA_HIERARCHY_NO_LABELS_NO_TYPE = """### Closed Issues without Pu
   - None: _HI350 open_ #350
     - _Release Notes_:
       - Sub-hierarchy level release note
-    - #453 _SI453 closed_ in #152
+    - N/A: #453 _SI453 closed_ in #152
       - Hierarchy level release note
       - Fixed bug
       - Improved performance
@@ -219,7 +218,7 @@ RELEASE_NOTES_DATA_HIERARCHY_WITH_LABELS_NO_TYPE = """### Epics
 - None: _HI302 open_ #302
   - _Release Notes_:
     - Hierarchy level release note
-  - #451 _SI451 closed_ in #150
+  - N/A: #451 _SI451 closed_ in #150
     - Hierarchy level release note
     - Fixed bug
     - Improved performance
@@ -228,7 +227,7 @@ RELEASE_NOTES_DATA_HIERARCHY_WITH_LABELS_NO_TYPE = """### Epics
 - None: _HI303 open_ #303
   - _Release Notes_:
     - Hierarchy level release note
-  - #452 _SI452 closed_ in #151
+  - N/A: #452 _SI452 closed_ in #151
     - Hierarchy level release note
     - Fixed bug
     - Improved performance
@@ -240,7 +239,7 @@ RELEASE_NOTES_DATA_HIERARCHY_WITH_LABELS_NO_TYPE = """### Epics
   - None: _HI350 open_ #350
     - _Release Notes_:
       - Sub-hierarchy level release note
-    - #453 _SI453 closed_ in #152
+    - N/A: #453 _SI453 closed_ in #152
       - Hierarchy level release note
       - Fixed bug
       - Improved performance
@@ -255,7 +254,7 @@ RELEASE_NOTES_DATA_HIERARCHY_WITH_LABELS_NO_TYPE = """### Epics
     * Awesome architecture
 
 ### Closed Issues without Pull Request ⚠️
-- #121 _Fix the bug_ in 
+- N/A: #121 _Fix the bug_ in 
   - Solo issue release note
 
 ### Closed Issues without User Defined Labels ⚠️
@@ -273,7 +272,7 @@ All closed PRs are linked to issues.
 - 🔔 None: _HI302 open_ #302
   - _Release Notes_:
     - Hierarchy level release note
-  - 🔔 #451 _SI451 closed_ in #150
+  - 🔔 N/A: #451 _SI451 closed_ in #150
     - Hierarchy level release note
     - Fixed bug
     - Improved performance
@@ -282,7 +281,7 @@ All closed PRs are linked to issues.
 - 🔔 None: _HI303 open_ #303
   - _Release Notes_:
     - Hierarchy level release note
-  - 🔔 #452 _SI452 closed_ in #151
+  - 🔔 N/A: #452 _SI452 closed_ in #151
     - Hierarchy level release note
     - Fixed bug
     - Improved performance
@@ -294,7 +293,7 @@ All closed PRs are linked to issues.
   - 🔔 None: _HI350 open_ #350
     - _Release Notes_:
       - Sub-hierarchy level release note
-    - 🔔 #453 _SI453 closed_ in #152
+    - 🔔 N/A: #453 _SI453 closed_ in #152
       - Hierarchy level release note
       - Fixed bug
       - Improved performance
@@ -315,7 +314,7 @@ RELEASE_NOTES_DATA_HIERARCHY_NO_LABELS_WITH_TYPE = """### Epics
 - Epic: _HI302 open_ #302
   - _Release Notes_:
     - Hierarchy level release note
-  - #451 _SI451 closed_ in #150
+  - Task: #451 _SI451 closed_ in #150
     - Hierarchy level release note
     - Fixed bug
     - Improved performance
@@ -324,7 +323,7 @@ RELEASE_NOTES_DATA_HIERARCHY_NO_LABELS_WITH_TYPE = """### Epics
 - Epic: _HI303 open_ #303
   - _Release Notes_:
     - Hierarchy level release note
-  - #452 _SI452 closed_ in #151
+  - Task: #452 _SI452 closed_ in #151
     - Hierarchy level release note
     - Fixed bug
     - Improved performance
@@ -336,7 +335,7 @@ RELEASE_NOTES_DATA_HIERARCHY_NO_LABELS_WITH_TYPE = """### Epics
   - Feature: _HI350 open_ #350
     - _Release Notes_:
       - Sub-hierarchy level release note
-    - #453 _SI453 closed_ in #152
+    - Task: #453 _SI453 closed_ in #152
       - Hierarchy level release note
       - Fixed bug
       - Improved performance
@@ -344,7 +343,7 @@ RELEASE_NOTES_DATA_HIERARCHY_NO_LABELS_WITH_TYPE = """### Epics
         * Awesome architecture
 
 ### Closed Issues without Pull Request ⚠️
-- #121 _Fix the bug_ in 
+- Feature: #121 _Fix the bug_ in 
   - Solo issue release note
 
 ### Closed Issues without User Defined Labels ⚠️
@@ -366,7 +365,7 @@ All closed issues contain at least one of user defined labels.
 - 🔔 Epic: _HI302 open_ #302
   - _Release Notes_:
     - Hierarchy level release note
-  - 🔔 #451 _SI451 closed_ in #150
+  - 🔔 Task: #451 _SI451 closed_ in #150
     - Hierarchy level release note
     - Fixed bug
     - Improved performance
@@ -375,7 +374,7 @@ All closed issues contain at least one of user defined labels.
 - 🔔 Epic: _HI303 open_ #303
   - _Release Notes_:
     - Hierarchy level release note
-  - 🔔 #452 _SI452 closed_ in #151
+  - 🔔 Task: #452 _SI452 closed_ in #151
     - Hierarchy level release note
     - Fixed bug
     - Improved performance
@@ -387,7 +386,7 @@ All closed issues contain at least one of user defined labels.
   - 🔔 Feature: _HI350 open_ #350
     - _Release Notes_:
       - Sub-hierarchy level release note
-    - 🔔 #453 _SI453 closed_ in #152
+    - 🔔 Task: #453 _SI453 closed_ in #152
       - Hierarchy level release note
       - Fixed bug
       - Improved performance
@@ -408,7 +407,7 @@ RELEASE_NOTES_DATA_HIERARCHY_WITH_LABELS_WITH_TYPE = """### Epics
 - Epic: _HI302 open_ #302
   - _Release Notes_:
     - Hierarchy level release note
-  - #451 _SI451 closed_ in #150
+  - Task: #451 _SI451 closed_ in #150
     - Hierarchy level release note
     - Fixed bug
     - Improved performance
@@ -417,7 +416,7 @@ RELEASE_NOTES_DATA_HIERARCHY_WITH_LABELS_WITH_TYPE = """### Epics
 - Epic: _HI303 open_ #303
   - _Release Notes_:
     - Hierarchy level release note
-  - #452 _SI452 closed_ in #151
+  - Task: #452 _SI452 closed_ in #151
     - Hierarchy level release note
     - Fixed bug
     - Improved performance
@@ -429,7 +428,7 @@ RELEASE_NOTES_DATA_HIERARCHY_WITH_LABELS_WITH_TYPE = """### Epics
   - Feature: _HI350 open_ #350
     - _Release Notes_:
       - Sub-hierarchy level release note
-    - #453 _SI453 closed_ in #152
+    - Task: #453 _SI453 closed_ in #152
       - Hierarchy level release note
       - Fixed bug
       - Improved performance
@@ -444,7 +443,7 @@ RELEASE_NOTES_DATA_HIERARCHY_WITH_LABELS_WITH_TYPE = """### Epics
     * Awesome architecture
 
 ### Closed Issues without Pull Request ⚠️
-- #121 _Fix the bug_ in 
+- Bug: #121 _Fix the bug_ in 
   - Solo issue release note
 
 ### Closed Issues without User Defined Labels ⚠️
@@ -462,7 +461,7 @@ All closed PRs are linked to issues.
 - 🔔 Epic: _HI302 open_ #302
   - _Release Notes_:
     - Hierarchy level release note
-  - 🔔 #451 _SI451 closed_ in #150
+  - 🔔 Task: #451 _SI451 closed_ in #150
     - Hierarchy level release note
     - Fixed bug
     - Improved performance
@@ -471,7 +470,7 @@ All closed PRs are linked to issues.
 - 🔔 Epic: _HI303 open_ #303
   - _Release Notes_:
     - Hierarchy level release note
-  - 🔔 #452 _SI452 closed_ in #151
+  - 🔔 Task: #452 _SI452 closed_ in #151
     - Hierarchy level release note
     - Fixed bug
     - Improved performance
@@ -483,7 +482,7 @@ All closed PRs are linked to issues.
   - 🔔 Feature: _HI350 open_ #350
     - _Release Notes_:
       - Sub-hierarchy level release note
-    - 🔔 #453 _SI453 closed_ in #152
+    - 🔔 Task: #453 _SI453 closed_ in #152
       - Hierarchy level release note
       - Fixed bug
       - Improved performance
@@ -502,15 +501,15 @@ http://example.com/changelog
 
 
 RELEASE_NOTES_NO_DATA_HIERARCHY_NO_LABELS_NO_TYPE = """### Closed Issues without Pull Request ⚠️
-- #121 _Fix the bug_ in 
+- N/A: #121 _Fix the bug_ in 
   - Solo issue release note
 - #450 _SI450 closed_ in 
   - Hierarchy level release note
 
 ### Closed Issues without User Defined Labels ⚠️
-- 🔔 #121 _Fix the bug_ in 
+- 🔔 N/A: #121 _Fix the bug_ in 
   - Solo issue release note
-- 🔔 #450 _SI450 closed_ in 
+- 🔔 Task: #450 _SI450 closed_ in 
   - Hierarchy level release note
 - #451 _SI451 closed_ in #150
   - Hierarchy level release note
@@ -743,7 +742,7 @@ http://example.com/changelog
 """
 
 RELEASE_NOTES_DATA_CUSTOM_CHAPTERS_MORE_LABELS_DUPLICITY_REDUCTION_ON = """### Chapter 1 🛠
-- #122 _I1+bug-enhancement_ in #101, #102
+- N/A: #122 _I1+bug-enhancement_ in #101, #102
   - Fixed bug
   - Improved performance
   + More nice code
@@ -771,10 +770,10 @@ http://example.com/changelog
 """
 
 RELEASE_NOTES_DATA_SERVICE_CHAPTERS_CLOSED_ISSUE_NO_PR_NO_USER_LABELS = """### Closed Issues without Pull Request ⚠️
-- #121 _Fix the bug_ in
+- N/A: #121 _Fix the bug_ in
 
 ### Closed Issues without User Defined Labels ⚠️
-- 🔔 #121 _Fix the bug_ in
+- 🔔 N/A: #121 _Fix the bug_ in
 
 #### Full Changelog
 http://example.com/changelog
@@ -804,7 +803,7 @@ http://example.com/changelog
 RELEASE_NOTES_DATA_SERVICE_CHAPTERS_CLOSED_PR_NO_ISSUE_SKIP_USER_LABELS = RELEASE_NOTES_NO_DATA_NO_WARNING_NO_EMPTY_CHAPTERS
 
 RELEASE_NOTES_DATA_SERVICE_CHAPTERS_OPEN_ISSUE_AND_MERGED_PR_NO_USER_LABELS = """### Merged PRs Linked to 'Not Closed' Issue ⚠️
-- #122 _I1 open_ in #101, #102
+- N/A: #122 _I1 open_ in #101, #102
   - PR 101 1st release note
   - PR 101 2nd release note
   - PR 102 1st release note
@@ -826,14 +825,14 @@ http://example.com/changelog
 """
 
 RELEASE_NOTES_DATA_CLOSED_ISSUE_NO_PR_WITH_USER_LABELS = """### Closed Issues without Pull Request ⚠️
-- #121 _Fix the bug_ in
+- N/A: #121 _Fix the bug_ in
 
 #### Full Changelog
 http://example.com/changelog
 """
 
 RELEASE_NOTES_DATA_CLOSED_ISSUE_WITH_PR_WITHOUT_USER_LABELS = """### Closed Issues without User Defined Labels ⚠️
-- #122 _I1_ in #101, #102
+- N/A: #122 _I1_ in #101, #102
   - Fixed bug
   - Improved performance
   + More nice code
@@ -866,10 +865,10 @@ http://example.com/changelog
 """
 
 RELEASE_NOTES_DATA_MERGED_PRS_WITH_OPEN_ISSUES = """### Merged PRs Linked to 'Not Closed' Issue ⚠️
-- #122 _I1 open_ in #101
+- N/A: #122 _I1 open_ in #101
   - PR 101 1st release note
   - PR 101 2nd release note
-- #123 _I2 open_ in #102
+- N/A: #123 _I2 open_ in #102
   - PR 102 1st release note
   - PR 102 2nd release note
 
@@ -878,7 +877,7 @@ http://example.com/changelog
 """
 
 RELEASE_NOTES_DATA_CLOSED_ISSUE_WITH_MERGED_PRS_WITHOUT_USER_LABELS = """### Closed Issues without User Defined Labels ⚠️
-- #121 _Fix the bug_ in #123
+- N/A: #121 _Fix the bug_ in #123
   - Fixed bug
   - Improved performance
   + More nice code
@@ -891,7 +890,7 @@ http://example.com/changelog
 RELEASE_NOTES_DATA_CLOSED_ISSUE_WITH_MERGED_PRS_WITH_USER_LABELS_WITH_SKIP_LABEL = RELEASE_NOTES_NO_DATA_NO_WARNING_NO_EMPTY_CHAPTERS
 
 RELEASE_NOTES_DATA_CLOSED_ISSUE_WITH_MERGED_PRS_WITH_USER_LABELS = """### Chapter 1 🛠
-- #122 _I1+bug_ in #124
+- N/A: #122 _I1+bug_ in #124
   - Fixed bug
   - Improved performance
   + More nice code
@@ -1603,7 +1602,7 @@ def test_build_hierarchy_rls_notes_no_labels_no_type(
     mock_rate_limit.rate.reset.timestamp.return_value = time.time() + 3600
     mock_github_client.get_rate_limit.return_value = mock_rate_limit
 
-    factory = IssueHierarchyRecordFactory(github=mock_github_client, home_repository=mock_repo)
+    factory = DefaultRecordFactory(github=mock_github_client, home_repository=mock_repo)
     records = factory.generate(mined_data_isolated_record_types_no_labels_no_type_defined)
 
     builder = ReleaseNotesBuilder(
@@ -1635,7 +1634,7 @@ def test_build_hierarchy_rls_notes_with_labels_no_type(
     mock_rate_limit.rate.reset.timestamp.return_value = time.time() + 3600
     mock_github_client.get_rate_limit.return_value = mock_rate_limit
 
-    factory = IssueHierarchyRecordFactory(github=mock_github_client, home_repository=mock_repo)
+    factory = DefaultRecordFactory(github=mock_github_client, home_repository=mock_repo)
     records = factory.generate(mined_data_isolated_record_types_with_labels_no_type_defined)
 
     builder = ReleaseNotesBuilder(
@@ -1667,7 +1666,7 @@ def test_build_hierarchy_rls_notes_no_labels_with_type(
     mock_rate_limit.rate.reset.timestamp.return_value = time.time() + 3600
     mock_github_client.get_rate_limit.return_value = mock_rate_limit
 
-    factory = IssueHierarchyRecordFactory(github=mock_github_client, home_repository=mock_repo)
+    factory = DefaultRecordFactory(github=mock_github_client, home_repository=mock_repo)
     records = factory.generate(mined_data_isolated_record_types_no_labels_with_type_defined)
 
     builder = ReleaseNotesBuilder(
@@ -1690,132 +1689,6 @@ def test_build_hierarchy_rls_notes_with_labels_with_type(
     mocker.patch("release_notes_generator.builder.builder.ActionInputs.get_print_empty_chapters", return_value=True)
     mocker.patch("release_notes_generator.builder.builder.ActionInputs.get_hierarchy", return_value=True)
     mocker.patch("release_notes_generator.builder.builder.ActionInputs.get_row_format_hierarchy_issue", return_value="{type}: _{title}_ {number}")
-
-    mock_github_client = mocker.Mock(spec=Github)
-
-    mock_rate_limit = mocker.Mock()
-    mock_rate_limit.rate.remaining = 10
-    mock_rate_limit.rate.reset.timestamp.return_value = time.time() + 3600
-    mock_github_client.get_rate_limit.return_value = mock_rate_limit
-
-    factory = IssueHierarchyRecordFactory(github=mock_github_client, home_repository=mock_repo)
-    records = factory.generate(mined_data_isolated_record_types_with_labels_with_type_defined)
-
-    builder = ReleaseNotesBuilder(
-        records=records,
-        changelog_url=DEFAULT_CHANGELOG_URL,
-        custom_chapters=custom_chapters_not_print_empty_chapters,
-    )
-
-    actual_release_notes = builder.build()
-
-    assert expected_release_notes == actual_release_notes
-
-
-def test_build_no_hierarchy_rls_notes_no_labels_no_type_with_hierarchy_data(
-        mocker, mock_repo,
-        custom_chapters_not_print_empty_chapters, mined_data_isolated_record_types_no_labels_no_type_defined
-):
-    expected_release_notes = RELEASE_NOTES_NO_DATA_HIERARCHY_NO_LABELS_NO_TYPE
-
-    mocker.patch("release_notes_generator.record.factory.default_record_factory.safe_call_decorator", side_effect=mock_safe_call_decorator)
-    mocker.patch("release_notes_generator.builder.builder.ActionInputs.get_print_empty_chapters", return_value=True)
-    mocker.patch("release_notes_generator.builder.builder.ActionInputs.get_hierarchy", return_value=False)
-    # mocker.patch("release_notes_generator.builder.builder.ActionInputs.get_row_format_hierarchy_issue", return_value="{type}: _{title}_ {number}")
-
-    mock_github_client = mocker.Mock(spec=Github)
-
-    mock_rate_limit = mocker.Mock()
-    mock_rate_limit.rate.remaining = 10
-    mock_rate_limit.rate.reset.timestamp.return_value = time.time() + 3600
-    mock_github_client.get_rate_limit.return_value = mock_rate_limit
-
-    factory = DefaultRecordFactory(github=mock_github_client, home_repository=mock_repo)
-    records = factory.generate(mined_data_isolated_record_types_no_labels_no_type_defined)
-
-    builder = ReleaseNotesBuilder(
-        records=records,
-        changelog_url=DEFAULT_CHANGELOG_URL,
-        custom_chapters=custom_chapters_not_print_empty_chapters,
-    )
-
-    actual_release_notes = builder.build()
-
-    assert expected_release_notes == actual_release_notes
-
-def test_build_no_hierarchy_rls_notes_with_labels_no_type_with_hierarchy_data(
-        mocker, mock_repo,
-        custom_chapters_not_print_empty_chapters, mined_data_isolated_record_types_with_labels_no_type_defined
-):
-    expected_release_notes = RELEASE_NOTES_NO_DATA_HIERARCHY_WITH_LABELS_NO_TYPE
-
-    mocker.patch("release_notes_generator.record.factory.default_record_factory.safe_call_decorator", side_effect=mock_safe_call_decorator)
-    mocker.patch("release_notes_generator.builder.builder.ActionInputs.get_print_empty_chapters", return_value=True)
-    mocker.patch("release_notes_generator.builder.builder.ActionInputs.get_hierarchy", return_value=False)
-
-    mock_github_client = mocker.Mock(spec=Github)
-
-    mock_rate_limit = mocker.Mock()
-    mock_rate_limit.rate.remaining = 10
-    mock_rate_limit.rate.reset.timestamp.return_value = time.time() + 3600
-    mock_github_client.get_rate_limit.return_value = mock_rate_limit
-
-    factory = DefaultRecordFactory(github=mock_github_client, home_repository=mock_repo)
-    records = factory.generate(mined_data_isolated_record_types_with_labels_no_type_defined)
-
-    builder = ReleaseNotesBuilder(
-        records=records,
-        changelog_url=DEFAULT_CHANGELOG_URL,
-        custom_chapters=custom_chapters_not_print_empty_chapters,
-    )
-
-    actual_release_notes = builder.build()
-
-    assert expected_release_notes == actual_release_notes
-
-
-def test_build_no_hierarchy_rls_notes_no_labels_with_type_with_hierarchy_data(
-        mocker, mock_repo,
-        custom_chapters_not_print_empty_chapters, mined_data_isolated_record_types_no_labels_with_type_defined
-):
-    expected_release_notes = RELEASE_NOTES_NO_DATA_HIERARCHY_NO_LABELS_WITH_TYPE
-
-    mocker.patch("release_notes_generator.record.factory.default_record_factory.safe_call_decorator", side_effect=mock_safe_call_decorator)
-    mocker.patch("release_notes_generator.builder.builder.ActionInputs.get_print_empty_chapters", return_value=True)
-    mocker.patch("release_notes_generator.builder.builder.ActionInputs.get_hierarchy", return_value=False)
-    # mocker.patch("release_notes_generator.builder.builder.ActionInputs.get_row_format_hierarchy_issue", return_value="{type}: _{title}_ {number}")
-
-    mock_github_client = mocker.Mock(spec=Github)
-
-    mock_rate_limit = mocker.Mock()
-    mock_rate_limit.rate.remaining = 10
-    mock_rate_limit.rate.reset.timestamp.return_value = time.time() + 3600
-    mock_github_client.get_rate_limit.return_value = mock_rate_limit
-
-    factory = DefaultRecordFactory(github=mock_github_client, home_repository=mock_repo)
-    records = factory.generate(mined_data_isolated_record_types_no_labels_with_type_defined)
-
-    builder = ReleaseNotesBuilder(
-        records=records,
-        changelog_url=DEFAULT_CHANGELOG_URL,
-        custom_chapters=custom_chapters_not_print_empty_chapters,
-    )
-
-    actual_release_notes = builder.build()
-
-    assert expected_release_notes == actual_release_notes
-
-
-def test_build_no_hierarchy_rls_notes_with_labels_with_type_with_hierarchy_data(
-        mocker, mock_repo,
-        custom_chapters_not_print_empty_chapters, mined_data_isolated_record_types_with_labels_with_type_defined
-):
-    expected_release_notes = RELEASE_NOTES_NO_DATA_HIERARCHY_WITH_LABELS_WITH_TYPE
-
-    mocker.patch("release_notes_generator.record.factory.default_record_factory.safe_call_decorator", side_effect=mock_safe_call_decorator)
-    mocker.patch("release_notes_generator.builder.builder.ActionInputs.get_print_empty_chapters", return_value=True)
-    mocker.patch("release_notes_generator.builder.builder.ActionInputs.get_hierarchy", return_value=False)
-    # mocker.patch("release_notes_generator.builder.builder.ActionInputs.get_row_format_hierarchy_issue", return_value="{type}: _{title}_ {number}")
 
     mock_github_client = mocker.Mock(spec=Github)
