+++ tests/test_release_notes_generator.py
@@ -78,8 +78,8 @@ def test_generate_release_notes_latest_release_not_found(
     release_notes = ReleaseNotesGenerator(github_mock, custom_chapters).generate()
 
     assert release_notes is not None
-    assert "- #121 _Fix the bug_" in release_notes
-    assert "- #122 _I1+bug_" in release_notes
+    assert "- N/A: #121 _Fix the bug_" in release_notes
+    assert "- N/A: #122 _I1+bug_" in release_notes
     assert "- PR: #101 _Fixed bug_" in release_notes
     assert "- PR: #102 _Fixed bug_" in release_notes
 
@@ -132,7 +132,7 @@ def test_generate_release_notes_latest_release_found_by_created_at(
     release_notes = ReleaseNotesGenerator(github_mock, custom_chapters).generate()
 
     assert release_notes is not None
-    assert "- #122 _I1+bug_" in release_notes
+    assert "- N/A: #122 _I1+bug_" in release_notes
     assert "- PR: #101 _Fixed bug_" not in release_notes
     assert "- PR: #102 _Fixed bug_" in release_notes
 
@@ -182,6 +182,6 @@ def test_generate_release_notes_latest_release_found_by_published_at(
     release_notes = ReleaseNotesGenerator(github_mock, custom_chapters).generate()
 
     assert release_notes is not None
-    assert "- #122 _I1+bug_" in release_notes
+    assert "- N/A: #122 _I1+bug_" in release_notes
     assert "- PR: #101 _Fixed bug_" not in release_notes
     assert "- PR: #102 _Fixed bug_" in release_notes
